+++ src/main/java/chess/domain/MoveRule/AbstractRule.java
@@ -1,4 +1,4 @@
-package chess.domain.RuleImpl;
+package chess.domain.MoveRule;
 
 import chess.domain.Position;
