+++ src/main/java/chess/domain/MoveRule/Bishop.java
@@ -1,4 +1,4 @@
-package chess.domain.RuleImpl;
+package chess.domain.MoveRule;
 
 import chess.domain.Position;
