+++ src/main/java/chess/dto/CommandDto.java
@@ -2,7 +2,7 @@ package chess.dto;
 
 public class CommandDto {
 	private long round;
-	private long room_id;
+	private long roomId;
 	private String origin;
 	private String target;
 
@@ -14,8 +14,8 @@ public class CommandDto {
 		this.origin = origin;
 	}
 
-	public void setRoomId(final long room_id) {
-		this.room_id = room_id;
+	public long getRoomId() {
+		return roomId;
 	}
 
 	public void setRound(final long round) {
@@ -26,8 +26,8 @@ public class CommandDto {
 		return round;
 	}
 
-	public long getRoom_id() {
-		return room_id;
+	public void setRoomId(final long room_id) {
+		this.roomId = room_id;
 	}
 
 	public String getOrigin() {
